fixed ray tests
